fix(fuse): isolate kernel cache bypass sweeper
